requirements.txt
# Flask and extensions
flask==3.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
werkzeug==3.0.1

# Database
sqlalchemy==2.0.23

# Telegram User API
telethon==1.34.0
cryptg==0.4.0

# Async support
aiofiles==23.2.1

# Utils
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1

# Security
bcrypt==4.1.2
cryptography==41.0.7

# Gunicorn for production
gunicorn==21.2.0

# Telegram Bot
python-telegram-bot==21.0



